Refactor model loading and support Real layer
Also improve usability of model loading (now only requires calling
`load_model`).
